requirements.txt
cleanlab
click
pandas
numba
catboost
openpyxl
dirhash
onnx
onnxruntime-gpu
hydra-core
altair
torch==1.10.1+cu113 
torchvision==0.11.2+cu113 
torchaudio==0.10.1+cu113
skorch
torchvision
pytorch-lightning
autogluon
annoy
cifar2png
tensorflow
tensorflow_datasets
umap-learn
seaborn
